+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## Devel
+
+- Bugfix: require records with `access=OWNER` to have `can_compute=True` for `WorkspaceGroupSharing` objects.
+
 ## 0.12.1 (2023-02-16)
 
 - Add form Mixin class that adds select2-bootstrap-5-theme to form media.